CORE-1495 Add Unit test for AuthController
