--- test/unit/com/unifina/controller/security/LoginControllerSpec.groovy
@@ -1,20 +1,0 @@
-package com.unifina.controller.security
-
-import grails.test.mixin.TestFor
-import spock.lang.Specification
-
-/**
- * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
- */
-@TestFor(LoginController)
-class LoginControllerSpec extends Specification {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-    }
-}
