--- test/unit/streamr/core/AuthControllerSpec.groovy
@@ -1,21 +1,0 @@
-package streamr.core
-
-import com.unifina.controller.security.AuthController
-import grails.test.mixin.TestFor
-import spock.lang.Specification
-
-/**
- * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
- */
-@TestFor(AuthController)
-class AuthControllerSpec extends Specification {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-    }
-}
